+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryConfig.java
@@ -81,6 +81,8 @@ public interface BigQueryConfig {
 
   Priority getQueryJobPriority();
 
+  long getBigQueryJobTimeoutInMinutes();
+
   default int getClientCreationHashCode() {
     return Objects.hashCode(
         getAccessTokenProviderFQCN(),